+++ src/components/lists/VcsList.vue
@@ -104,6 +104,7 @@
           <vcs-action-button-list
             v-if="item.actions?.length > 0"
             :actions="item.actions"
+            :disabled="item.disabled"
             :block-overflow="true"
             :overflow-count="actionButtonListOverflowCount"
           />